feat: create title screen at the start

+++ src/client/components/RenderComponent.java
@@ -6,16 +6,16 @@ import org.joml.Vector4f;
 
 public class RenderComponent implements Component {
     public Texture texture;
-    public int z;
+    public float z;
     public Vector4f tint;
 
-    public RenderComponent(Texture texture, int z, Vector4f tint) {
+    public RenderComponent(Texture texture, float z, Vector4f tint) {
         this.texture = texture;
         this.z = z;
         this.tint = tint;
     }
 
-    public RenderComponent(Texture texture, int z) {
+    public RenderComponent(Texture texture, float z) {
         this(texture, z, new Vector4f(1.0f, 1.0f, 1.0f, 1.0f));
     }
